database schema cleaned up w/ alan's notes, and dummy data script added to migrations

+++ db/migrations/20190310200712_create-passengers-table.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   CREATE TABLE Passengers (
 	uid 			INTEGER NOT NULL REFERENCES Users(uid),
 	tripsTaken		INTEGER NOT NULL,
-	rating			FLOAT,
+	rating			NUMERIC,
 	PRIMARY KEY (uid)	
 )`;
   return knex.raw(createQuery);
